3.1.2

+++ src/main/java/ru/kata/spring/boot_security/demo/Service/UserService.java
@@ -8,8 +8,8 @@ import java.util.List;
 
 public interface UserService {
     List<User> index();
-    User show(int id);
+    User show(long id);
     void save(User user);
     void update(User updatedUser);
-    void delete(int id);
+    void delete(long id);
 }
