+++ src/main/java/ru/kata/spring/boot_security/demo/Service/UserServiceImp.java
@@ -24,7 +24,7 @@ public class UserServiceImp implements UserService{
 
     @Override
     @Transactional(readOnly = true)
-    public User show(int id) {
+    public User show(long id) {
         return userDao.show(id);
     }
 
@@ -42,7 +42,7 @@ public class UserServiceImp implements UserService{
 
     @Override
     @Transactional
-    public void delete(int id) {
+    public void delete(long id) {
         userDao.delete(id);
     }
 }
