+++ src/main/java/ru/kata/spring/boot_security/demo/dao/UserDao.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface UserDao {
     List<User> index();
-    User show(int id);
+    User show(long id);
     void save(User user);
     void update(User updatedUser);
-    void delete(int id);
+    void delete(long id);
 
 }
